Smart stored procedure
This is an script to create a Job that uses Olla`s  IndexOptmize. But I added a step to check if there is enough space in the Lod drive and a stored procedure that accepts 2 parameters, lower size and high size. So if you pass 0 and 5 it will start the maintenance to all dbs that have between 0 and 5 GB.

Please adjust accordingly.